+++ src/lib/components/SheetToAddPayments.svelte
@@ -58,31 +58,31 @@
 			<div class="flex gap-4">
 				<div class="grid gap-2 w-full">
 					<Label for="name">Nombres</Label>
-					<Input class="text-lg" type="text" id="name" name="name" placeholder="Pablito" required />
+					<Input class="text-lg" type="text" id="name" name="name"  required />
 				</div>
 				<div class="grid gap-2 w-full">
 					<Label for="rut">RUT</Label>
-					<Input class="text-lg" type="text" id="rut" name="rut" placeholder="1111111-0" required />
+					<Input class="text-lg" type="text" id="rut" name="rut" />
 				</div>
 			</div>
 			<div class="flex gap-4">
 				<div class="grid gap-2 w-full">
 					<Label for="email">Email</Label>
-					<Input class="text-lg"  type="email" id="email" name="email" placeholder="pablito@5lc.cl" required />
+					<Input class="text-lg"  type="email" id="email" name="email" />
 				</div>
 				<div class="grid gap-2 w-full">
 					<Label for="phone">Teléfono</Label>
-					<Input class="text-lg" id="phone" name="phone" type="text" placeholder="+56991291468" required />
+					<Input class="text-lg" id="phone" name="phone" type="text"  />
 				</div>
 			</div>
 			<div class="flex gap-4">
 				<div class="grid gap-2 w-full">
 					<Label for="ticketAmount">Entradas</Label>
-					<Input class="text-lg" type="number" name="ticketAmount" min="0" max="10" id="ticketAmount" required />
+					<Input class="text-lg" type="number" name="ticketAmount" min="0" max="10" id="ticketAmount" value="1" required />
 				</div>
 				<div class="grid gap-2 w-full">
 					<Label for="price">Precio</Label>
-					<Input class="text-lg" type="number" name="price" min="0" id="price" required />
+					<Input class="text-lg" type="number" name="price" min="0" id="price" required value="0" />
 				</div>
 			</div>
 			{#if $page.data.eventFromSanityStudio?.sell_type === 'ubication'}