+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fix
 
 - display name in comments
+- adding new payments
+- Leave not required some inputs to add new payments
 
 ## [0.0.22] - 2025-01-13